Improved validation of profile existence in the Databricks configuration file
Previously, the verification of the existing profile in the Databricks configuration file was performed in the databricks_configure rule.
This was generating a bug, because when the first execution of the rule returned that a Profile did not exist, if the login was performed later, the result was always the non-existence of the profile.
The correction consisted of moving the check to the Run phase of the databricks_fs and databricks_libraries rules.

Some helpers have been moved to the directory where the rules code is.
Syntax and indentation adjustments were also considered.